Resolve warnings

+++ include/sparseir/basis_set.hpp
@@ -61,13 +61,13 @@ public:
         bool fence = false;
         bool positive_only = false;
         for (const auto &freq : basis_f->default_matsubara_sampling_points(
-                 basis_f->size(), fence, positive_only)) {
+                 static_cast<int>(basis_f->size()), fence, positive_only)) {
             wn_f.push_back(freq.get_n());
         }
 
         std::vector<int> wn_b;
         for (const auto &freq : basis_b->default_matsubara_sampling_points(
-                 basis_b->size(), fence, positive_only)) {
+                 static_cast<int>(basis_b->size()), fence, positive_only)) {
             wn_b.push_back(freq.get_n());
         }
